Scripts/sandbox/requirements.txt
boto3==1.4.5
botocore==1.5.91
certifi==2017.7.27.1
chardet==3.0.4
docutils==0.13.1
futures==3.1.1
idna==2.5
jmespath==0.9.3
pycrypto==2.6.1
python-dateutil==2.6.1
PyYAML==3.12
requests==2.18.2
s3transfer==0.1.10
six==1.10.0
slacker==0.9.50
urllib3==1.22
